Изменение: убрал warning;
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@2862 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/source/DTSAC3Source/DTSAC3Source.cpp
@@ -297,7 +297,7 @@ CDTSAC3Stream::CDTSAC3Stream(const WCHAR* wfn, CSource* pParent, HRESULT* phr)
 			m_dataEnd   = m_dataStart + datasize;
 		}
 
-		if (m_file.GetLength() < m_dataEnd) {
+		if ((__int64)m_file.GetLength() < m_dataEnd) {
 			m_dataEnd = m_file.GetLength();
 		}
